+++ src/App.jsx
@@ -38,6 +38,7 @@ const { name, version } = packageJson
 const defaultSettings = {
    version: version,
    dm: false,
+   gs: false,
    header: {
       gh: "https://github.com",
       li: "https://linkedin.com",
@@ -82,6 +83,7 @@ export default function App() {
    }
 
    const [darkMode, setDarkMode] = useState(ls.dm ?? false)
+   const [gridSet, setGridSet] = useState(ls.gs ?? false)
 
    useEffect(() => {
       if (darkMode) {
@@ -91,7 +93,7 @@ export default function App() {
       }
    }, [darkMode])
 
-   const themeProps = { name, ls, darkMode, setDarkMode }
+   const themeProps = { name, ls, darkMode, setDarkMode, gridSet, setGridSet }
    const theme = createTheme({
       palette: { mode: darkMode ? "dark" : "light" }
    })