+++ src/pages/settings/Settings.jsx
@@ -1,144 +1,342 @@
-import PropTypes from "prop-types"
-import { useState, useEffect } from 'react'
-import { useNavigate, useLocation } from "react-router-dom"
-import { Box, Container, Grid, Link } from "@mui/material"
-import { TextField, Button, Switch } from "@mui/material"
-import { styled } from '@mui/material/styles'
+import PropTypes from "prop-types";
+import { useState, useEffect } from "react";
+import { useTheme } from "@mui/material/styles";
+import { useNavigate, useLocation } from "react-router-dom";
+import { Box, Container, Grid, Link } from "@mui/material";
+import { TextField, Button, Switch, Tabs, Tab } from "@mui/material";
+import { styled } from "@mui/material/styles";
 
+/**
+ * The `Settings` component is responsible for rendering the settings page of the application.
+ * It allows the user to update various settings, such as their social media URLs and dark mode preferences.
+ * The component uses the `themeProps` object to access the current theme settings, including the dark mode state and a function to toggle it.
+ * The component also provides functionality to reset the settings to their default values and to navigate back to the home page.
+ */
+export default function Settings({ themeProps }) {
+   const { name, ls, darkMode, setDarkMode, gridSet, setGridSet } = themeProps;
 
+   const lcolor = darkMode ? "lightblue" : "#1b2051";
 
-export default function Settings({themeProps}) {
-   const { name, ls, darkMode, setDarkMode } = themeProps
-
-   const lcolor = (darkMode) ? "lightblue" : "#1b2051"
-
-   const Div = styled('div')(({ theme }) => ({
+   const Div = styled("div")(({ theme }) => ({
       ...theme.typography.button,
-      textTransform: 'none',
+      textTransform: "none",
       marginTop: "-1px",
       color: lcolor,
-      fontSize: "14pt",
-   }))
+      fontSize: "10pt",
+   }));
 
+   const { gh, li, tw, ig, fb, ic, ib, ca, go, sm, ai } = ls.header;
 
-   const {gh,li,tw,ig,fb,ic,ib,ca,go,sm,ai} = ls.header
-
-   const [githubUrl, setGithubUrl] = useState(gh ?? '')
-   const [linkedinUrl, setLinkedinUrl] = useState(li ??'')
-   const [twitterUrl, setTwitterUrl] = useState(tw ?? '')
-   const [instaUrl, setInstaUrl] = useState(ig ?? '')
-   const [facebookUrl, setFacebookUrl] = useState(fb ?? '')
-   const [icloudUrl, setIcloudUrl] = useState(ic ?? '')
-   const [inboxUrl, setInboxUrl] = useState(ib ?? '')
-   const [calendarUrl, setCalendarUrl] = useState(ca ?? '')
-   const [travelUrl, setTravelUrl] = useState(go ?? '')
-   const [brokerUrl, setBrokerUrl] = useState(sm ?? '')
-   const [chataiUrl, setChataiUrl] = useState(ai ?? '')
-
+   const [githubUrl, setGithubUrl] = useState(gh ?? "");
+   const [linkedinUrl, setLinkedinUrl] = useState(li ?? "");
+   const [twitterUrl, setTwitterUrl] = useState(tw ?? "");
+   const [instaUrl, setInstaUrl] = useState(ig ?? "");
+   const [facebookUrl, setFacebookUrl] = useState(fb ?? "");
+   const [icloudUrl, setIcloudUrl] = useState(ic ?? "");
+   const [inboxUrl, setInboxUrl] = useState(ib ?? "");
+   const [calendarUrl, setCalendarUrl] = useState(ca ?? "");
+   const [travelUrl, setTravelUrl] = useState(go ?? "");
+   const [brokerUrl, setBrokerUrl] = useState(sm ?? "");
+   const [chataiUrl, setChataiUrl] = useState(ai ?? "");
+   // Add state for active tab
+   const [value, setValue] = useState(0);
 
    useEffect(() => {
-      setGithubUrl(gh)
-      setLinkedinUrl(li)
-      setTwitterUrl(tw)
-      setInstaUrl(ig)
-      setFacebookUrl(fb)
-      setIcloudUrl(ic)
-      setInboxUrl(ib)
-      setCalendarUrl(ca)
-      setTravelUrl(go)
-      setBrokerUrl(sm)
-      setChataiUrl(ai)
-    }, [gh, li, tw, ig, fb, ic, ib, ca, go, sm, ai])
+      setGithubUrl(gh);
+      setLinkedinUrl(li);
+      setTwitterUrl(tw);
+      setInstaUrl(ig);
+      setFacebookUrl(fb);
+      setIcloudUrl(ic);
+      setInboxUrl(ib);
+      setCalendarUrl(ca);
+      setTravelUrl(go);
+      setBrokerUrl(sm);
+      setChataiUrl(ai);
+   }, [gh, li, tw, ig, fb, ic, ib, ca, go, sm, ai]);
 
+   const box1 = {
+      fontFamily: "Verdana, Geneva, Tahoma, sans-serif",
+      margin: "12px 0px 0px 2px",
+   };
+   const button1 = {
+      height: 18,
+      textTransform: "none",
+      marginLeft: "300px",
+      textAlign: "center",
+      marginTop: "6px",
+   };
+   const button2 = {
+      textTransform: "none",
+      paddingRight: 2,
+      fontSize: "10pt",
+      fontWeight: 600,
+      textAlign: "center",
+   };
+   const line = { margin: "8px 0px 8px 18px", width: "240px" };
+   const lline = { margin: "8px 0px 8px 18px", width: "240px" };
 
-    const box1 = { "fontFamily": "Verdana, Geneva, Tahoma, sans-serif", "margin": "12px 0px 0px 2px" }
-    const button1 = { height: 26, textTransform: "none", marginLeft: "300px", textAlign:"center", marginTop: "6px" }
-    const button2 = { textTransform: "none", paddingRight: 2, fontSize: "10pt", fontWeight:600, textAlign: "center"}
-    const line = {margin:"8px 0px 8px 18px",width:"240px"}
-    const lline = {margin:"8px 0px 8px 18px",width:"240px"}
+   const pagelink = "/settings";
+   const navigate = useNavigate();
+   const location = useLocation();
+   const handleNav = () => {
+      navigate(location.pathname === pagelink ? "/" : pagelink);
+   };
+   const handleCancel = () => {
+      handleNav();
+   };
 
-
-    const pagelink = "/settings"
-    const navigate = useNavigate()
-    const location = useLocation()
-    const handleNav = () => { navigate(location.pathname === pagelink ? "/" : pagelink) }
-    const handleCancel = () => { handleNav() }
-    const handleReset = () => { localStorage.removeItem(name);chrome.runtime.reload() }
-
-    const handleUpdate = () => {
-      ls.header.gh = githubUrl
-      ls.header.li = linkedinUrl
-      ls.header.tw = twitterUrl
-      ls.header.ig = instaUrl
-      ls.header.fb = facebookUrl
-      ls.header.ic = icloudUrl
-      ls.header.ib = inboxUrl
-      ls.header.ca = calendarUrl
-      ls.header.go = travelUrl
-      ls.header.sm = brokerUrl
-      ls.header.ai = chataiUrl
-      localStorage.setItem(name, JSON.stringify(ls))
-      handleNav() }
-
-       const toggleDarkMode = () => {
-         ls.dm = !darkMode
-         setDarkMode(ls.dm)
-         localStorage.setItem(name,JSON.stringify(ls))
+   /**
+    * Handles resetting the application state by removing the stored data from localStorage and reloading the page.
+    * This function is called when the user wants to reset their settings or preferences.
+    */
+   const handleReset = () => {
+      localStorage.removeItem(name);
+      // eslint-disable-next-line no-undef
+      if (typeof chrome !== "undefined" && chrome.runtime) {
+         // eslint-disable-next-line no-undef
+         chrome.runtime.reload();
+      } else {
+         window.location.reload();
       }
+   };
 
+   const handleUpdate = () => {
+      ls.header.gh = githubUrl;
+      ls.header.li = linkedinUrl;
+      ls.header.tw = twitterUrl;
+      ls.header.ig = instaUrl;
+      ls.header.fb = facebookUrl;
+      ls.header.ic = icloudUrl;
+      ls.header.ib = inboxUrl;
+      ls.header.ca = calendarUrl;
+      ls.header.go = travelUrl;
+      ls.header.sm = brokerUrl;
+      ls.header.ai = chataiUrl;
+      localStorage.setItem(name, JSON.stringify(ls));
+      handleNav();
+   };
 
-    return (
+   const toggleDarkMode = () => {
+      ls.dm = !darkMode;
+      setDarkMode(ls.dm);
+      localStorage.setItem(name, JSON.stringify(ls));
+   };
+   const toggleGridSet = () => {
+      ls.gs = !gridSet;
+      setGridSet(ls.gs);
+      localStorage.setItem(name, JSON.stringify(ls));
+   };
+
+   const handleChange = (event, newValue) => {
+      setValue(newValue);
+   };
+   const tabStyle = {
+      fontSize: "10pt",
+      textTransform: "none",
+   };
+
+   return (
       <Box sx={box1}>
-         <Container sx={{ margin: "18px 22px 18px 22px" }}>
-            <Div>Settings - <Link sx={{textShadow:"none", textDecoration:"none", color:lcolor,'&:hover':{color:lcolor} }} href="https://chrome.google.com/webstore/developer/dashboard">Gridlinks</Link></Div>
+         <Container sx={{ margin: "8px 8px 0px 0px" }}>
+            <Div>
+               Settings -{" "}
+               <Link
+                  sx={{
+                     textShadow: "none",
+                     textDecoration: "none",
+                     color: lcolor,
+                     "&:hover": { color: lcolor },
+                  }}
+                  href="https://chrome.google.com/webstore/developer/dashboard"
+               >
+                  Gridlinks
+               </Link>
+            </Div>
          </Container>
-         <form onSubmit={handleUpdate}>
 
-            <TextField name="github" key="githubUrl" label="GitHub URL" size="small" sx={line}
-               onChange={(event) => setGithubUrl(event.target.value)} value={githubUrl} />
+         {/* Tab control */}
+         <Tabs
+            value={value}
+            onChange={handleChange}
+            aria-label="settings tabs"
+            sx={{
+               marginLeft: "18px",
+               marginRight: "18px",
+            }}
+         >
+            <Tab label="General" sx={tabStyle} />
+            <Tab label="Icon Links" sx={tabStyle} />
+         </Tabs>
 
-            <TextField name="linkedin" key="linkedinUrl" label="LinkedIn URL" size="small" sx={line}
-               onChange={(event) => setLinkedinUrl(event.target.value)} value={linkedinUrl} />
+         {/* Conditional rendering based on the active tab */}
+         {value === 0 && (
+            <Box sx={{ marginLeft:"22px", height: "224px", fontSize: "10pt", padding: "8px" }}>
+               {/* Appearance settings or other content for the "Appearance" tab */}
+               Appearance:
+               <Switch
+                  sx={{ ...button2, marginLeft: "22px" }}
+                  checked={darkMode}
+                  onChange={toggleDarkMode}
+                  inputProps={{ "aria-label": "Dark mode toggle" }}
+               />
+               dark mode
+               <br/>
+               Gridset:
 
-            <TextField name="x.com" key="twitterUrl" label="X.com URL" size="small" sx={line}
-               onChange={(event) => setTwitterUrl(event.target.value)} value={twitterUrl} />
+               <Switch
+                  sx={{ ...button2, marginLeft: "54px" }}
+                  checked={gridSet}
+                  onChange={toggleGridSet}
+                  inputProps={{ "aria-label": "Alternate set toggle" }}
+               />
+               alternate
+            </Box>
+         )}
+         {value === 1 && (
+            <form onSubmit={handleUpdate}>
+               <TextField
+                  name="github"
+                  key="githubUrl"
+                  label="GitHub URL"
+                  size="small"
+                  sx={line}
+                  onChange={(event) => setGithubUrl(event.target.value)}
+                  value={githubUrl}
+               />
 
-            <TextField name="instagram" key="instaUrl" label="Instagram URL" size="small" sx={line}
-               onChange={(event) => setInstaUrl(event.target.value)} value={instaUrl} />
+               <TextField
+                  name="linkedin"
+                  key="linkedinUrl"
+                  label="LinkedIn URL"
+                  size="small"
+                  sx={line}
+                  onChange={(event) => setLinkedinUrl(event.target.value)}
+                  value={linkedinUrl}
+               />
 
-            <TextField name="facebook" key="facebookUrl" label="Facebook URL" size="small" sx={line}
-               onChange={(event) => setFacebookUrl(event.target.value)} value={facebookUrl} />
+               <TextField
+                  name="x.com"
+                  key="twitterUrl"
+                  label="X.com URL"
+                  size="small"
+                  sx={line}
+                  onChange={(event) => setTwitterUrl(event.target.value)}
+                  value={twitterUrl}
+               />
 
-            <TextField name="icloud" key="icloudUrl" label="iCloud URL" size="small" sx={line}
-               onChange={(event) => setIcloudUrl(event.target.value)} value={icloudUrl} />
+               <TextField
+                  name="instagram"
+                  key="instaUrl"
+                  label="Instagram URL"
+                  size="small"
+                  sx={line}
+                  onChange={(event) => setInstaUrl(event.target.value)}
+                  value={instaUrl}
+               />
 
-            <TextField name="inbox" key="inboxUrl" label="inbox URL" size="small" sx={line}
-               onChange={(event) => setInboxUrl(event.target.value)} value={inboxUrl} />
+               <TextField
+                  name="facebook"
+                  key="facebookUrl"
+                  label="Facebook URL"
+                  size="small"
+                  sx={line}
+                  onChange={(event) => setFacebookUrl(event.target.value)}
+                  value={facebookUrl}
+               />
 
-            <TextField name="calendar" key="calendarUrl" label="Calendar URL" size="small" sx={line}
-               onChange={(event) => calendarUrl(event.target.value)} value={calendarUrl} />
+               <TextField
+                  name="icloud"
+                  key="icloudUrl"
+                  label="iCloud URL"
+                  size="small"
+                  sx={line}
+                  onChange={(event) => setIcloudUrl(event.target.value)}
+                  value={icloudUrl}
+               />
 
-            <TextField name="travel" key="travelUrl" label="Travel URL" size="small" sx={line}
-               onChange={(event) => setTravelUrl(event.target.value)} value={travelUrl} />
+               <TextField
+                  name="inbox"
+                  key="inboxUrl"
+                  label="inbox URL"
+                  size="small"
+                  sx={line}
+                  onChange={(event) => setInboxUrl(event.target.value)}
+                  value={inboxUrl}
+               />
 
-            <TextField name="brokerage" key="brokerUrl" label="Broker URL" size="small" sx={line}
-               onChange={(event) => setBrokerUrl(event.target.value)} value={brokerUrl} />
+               <TextField
+                  name="calendar"
+                  key="calendarUrl"
+                  label="Calendar URL"
+                  size="small"
+                  sx={line}
+                  onChange={(event) => calendarUrl(event.target.value)}
+                  value={calendarUrl}
+               />
 
-            <TextField name="genai" key="chataiUrl" label="AI URL" size="small" sx={lline}
-               onChange={(event) => setChataiUrl(event.target.value)} value={chataiUrl} />
+               <TextField
+                  name="travel"
+                  key="travelUrl"
+                  label="Travel URL"
+                  size="small"
+                  sx={line}
+                  onChange={(event) => setTravelUrl(event.target.value)}
+                  value={travelUrl}
+               />
 
+               <TextField
+                  name="brokerage"
+                  key="brokerUrl"
+                  label="Broker URL"
+                  size="small"
+                  sx={line}
+                  onChange={(event) => setBrokerUrl(event.target.value)}
+                  value={brokerUrl}
+               />
 
+               <TextField
+                  name="genai"
+                  key="chataiUrl"
+                  label="AI URL"
+                  size="small"
+                  sx={lline}
+                  onChange={(event) => setChataiUrl(event.target.value)}
+                  value={chataiUrl}
+               />
+            </form>
+         )}
          <Grid item sx={button1}>
-            <Switch sx={{...button2, marginLeft:"22px"}}  checked={darkMode} onChange={toggleDarkMode} inputProps={{ 'aria-label': 'Dark mode toggle' }}/>
-            dark mode
-            <Button sx={{...button2, marginLeft:"22px"}} variant="contained" size="small" color="primary" onClick={handleReset}>Reset</Button>
-            <Button sx={{...button2, marginLeft:"22px"}} variant="contained" size="small" color="primary" onClick={handleCancel}>Cancel</Button>
-            <Button sx={{...button2, marginLeft:"22px"}} variant="contained" size="small" color="success" onClick={handleUpdate}>Update</Button>
+            <Button
+               sx={{ ...button2, marginLeft: "22px" }}
+               variant="contained"
+               size="small"
+               color="primary"
+               onClick={handleReset}
+            >
+               Reset
+            </Button>
+            <Button
+               sx={{ ...button2, marginLeft: "22px" }}
+               variant="contained"
+               size="small"
+               color="primary"
+               onClick={handleCancel}
+            >
+               Cancel
+            </Button>
+            <Button
+               sx={{ ...button2, marginLeft: "22px" }}
+               variant="contained"
+               size="small"
+               color="success"
+               onClick={handleUpdate}
+            >
+               Update
+            </Button>
          </Grid>
-
-      </form>
-   </Box>
-       )
+      </Box>
+   );
 }
 
 Settings.propTypes = {
@@ -147,5 +345,7 @@ Settings.propTypes = {
       ls: PropTypes.object.isRequired,
       darkMode: PropTypes.bool.isRequired,
       setDarkMode: PropTypes.func.isRequired,
+      gridSet: PropTypes.bool.isRequired,
+      setGridSet: PropTypes.func.isRequired,
    }).isRequired,
-}
+};